Overhaul, more flexible now
Learned more about "package unknown", and now the tinlist can be expanded, with multiple packages in one repo. Syntax extended to allow for -exact $version, and -force for "tin import". Removed "tin extract" and "tin provide". No longer needed, and the "tinstall.tcl" idea is obsolete. Just need to create an installer script for your repo, and it will work with tin.