Update tin.tin
Now if you specify a pkgUninstall.tcl file, it will run it just like a pkgIndex.tcl file in order to uninstall the package.
This is in case a package includes files outside the folder. By default, if no pkgUninstall.tcl file is available, it will just delete the folder. This also gives flexibility if there is user data in the package folder.